Refactor into single Tower class
The Tower class itself is the pillar object now and provides functions to
rendering engines. This can be used to `update`, `format` and `merge` to
the tower pillar data as well as to `include`, `lookup` or `load` additional
tower data files.

The merge code has been removed and rewritten to proper deepcopy merged
values to avoid mutating dicts inserted at different places (e.g. merged
default values). The new code does not yet has merge strategies.

The tower passes additional variables such as `tmplpath` and `tmpldir` to
renderers to match the behavior of the yamlet renderer.

Last but not least the commit adds more tests.