Add conf flag to enable saltenv passing to rendering engines
By setting the salt master conf `salt_tower.unstable_enable_saltenv`
to `True`, the salt tower ext_pillar will pass the `saltenv` to rendering
engines. Some will behave differently if `saltenv` is set, as it is when
rendering states or pillars, e.g. the JINJA renderer. It will lookup
files (e.g. in `import_yaml` or `include` macros) from the salt masters
`pillar_roots` directories, instead of relative to the current tower
data file.

Because this feature is hard to handle and depends on salts internal
behavior it is disabled by default. It might still be useful, e.g. if
the salt tower root also is a pillar_roots directory. If the feature
proves to be stable it might be enabled by default in the future.

If you only want to import relative files via JINJA in a tower data
file, you can use the `tmplpath`, `tmpldir` and `basedir` variables too:

    {% import_yaml tmpldir + "/data.yml" as data %}