Add all-sls include more to ease including directories
The all-sls include mode is a feature flag that changes the behavior how
files are loaded when an include statement matches a directory.

The default mode looks for a `dir/init.sls`, similar to how salt works.
The new `all-sls` will instead load all `*.sls` files, ordered by
filename. This can ease working with multiple renderers, or when
splitting up a role or module into multiple files. It avoid having to
modify the tower file to introduce globs for includes.