feat: support optional includes via YAML tag
This CL adds support for optional includes in the `tower.sls`. If they
are not found and raising an error on missing files is enabled, these
include still be ignored. The intended use case are broad includes in
`tower.sls` such as:

    base:
      # broad grain matchers to load some shared defaults for e.g.
      # some operating systems
      - !optional dist/{{ grains["oscodename"] }}/*.sls

      # specific minion overrides when needed
      - !optional minion/{{ minion_id }}

Having errors on missing includes otherwise avoids possible issues in
your pillar, and catches forgotten parts and removals.

The Yamlet renderer has been extended to take a secret argument for
declaring additional YAML tags and constructor functions. This allows
the tower to inject the `!optional` tag only when rendering the top
`tower.sls`. This feature is currently not intended for any other use.

Note: This means that `!optional` will only be supported when the yamlet
render is used for `tower.sls`.