feat: Raise error on missing file
When a file is included via `!include` or `!read` tags that does not
exist the renderer will now raise an error. Previously, the renderer
passed the file path to salt, that only returned a default dict and
emitted a log message.

Raising an error will abort compiling the pillar and pass up the error
up to the minion. This should catch most cases of missing or wrong
included paths and not return unexpected empty/blank pillars.